Fix warning when compiling with OpenSSL.

+++ vio/viossl.c
@@ -333,7 +333,7 @@ static int ssl_do(struct st_VioSSLFd *ptr, Vio *vio, long timeout,
   DBUG_PRINT("info", ("ssl: %p timeout: %ld", ssl, timeout));
   SSL_clear(ssl);
   SSL_SESSION_set_timeout(SSL_get_session(ssl), timeout);
-  SSL_set_fd(ssl, sd);
+  SSL_set_fd(ssl, (int)sd);
 
   /*
     Since yaSSL does not support non-blocking send operations, use
